Removed initializer from interface to base class.
